chore: create buttons base definitions
