+++ packages/accessibility-buttons/src/A11yFontSizeDecreaseButton/a11y-font-size-decrease-button.ts
@@ -1,0 +1,3 @@
+import { A11yFontDecreaseButton } from './A11yFontSizeDecreaseButton'
+
+globalThis.customElements.define('a11y-font-size-decrease-button', A11yFontDecreaseButton)
