+++ packages/accessibility-buttons/src/A11yFontSizeDecreaseButton/index.ts
@@ -1,0 +1,1 @@
+export { A11yFontDecreaseButton } from './A11yFontSizeDecreaseButton'
