+++ packages/accessibility-buttons/src/A11yFontSizeIncreaseButton/a11y-font-size-increase-button.ts
@@ -1,0 +1,3 @@
+import { A11yFontIncreaseButton } from './A11yFontSizeIncreaseButton'
+
+globalThis.customElements.define('a11y-font-size-increase-button', A11yFontIncreaseButton)
