+++ packages/accessibility-buttons/src/A11yFontSizeIncreaseButton/index.ts
@@ -1,0 +1,1 @@
+export { A11yFontIncreaseButton } from './A11yFontSizeIncreaseButton'
