+++ packages/accessibility-buttons/src/index.ts
@@ -1,6 +1,5 @@
 export * from './A11yContrastButton'
 export * from './A11yFontSizeButton'
-export * from './A11yFontSteps'
 
-export * from './A11yIncreaseButton/a11y-font-increase-button'
-export * from './A11yDecreaseButton/a11y-font-decrease-button'
+export * from './A11yFontSizeDecreaseButton'
+export * from './A11yFontSizeIncreaseButton'
