+++ packages/accessibility-buttons/src/utils/font-steps.ts
@@ -1,4 +1,5 @@
 export const DEFAULT_STEPS = 3
+
 const BASE_ATTR = 'data-a11y-font-base-size'
 const STEP_ATTR = 'data-a11y-font-step'
 const PER_STEP = 0.1 // 10% por etapa
@@ -13,7 +14,7 @@ export function applyStep(step: number) {
   document_.setAttribute(STEP_ATTR, String(step))
 
   let base = Number.parseFloat(document_.getAttribute(BASE_ATTR) || '')
-  if (!base || isNaN(base)) {
+  if (!base || Number.isNaN(base)) {
     base = Number.parseFloat(getComputedStyle(document_).fontSize) || 16
     document_.setAttribute(BASE_ATTR, String(base))
   }
