+++ src/index.html
@@ -213,6 +213,16 @@
     </header>
 
     <main class="main">
+      <a11y-font-size-increase-button
+        steps="3"
+        label="+A"
+      ></a11y-font-size-increase-button>
+
+      <a11y-font-size-decrease-button
+        steps="3"
+        label="-A"
+      ></a11y-font-size-decrease-button>
+
       <div class="row">
         <a11y-contrast-button
           add-contrast-label="Click me and things will be darken"
