+++ src/styles/button.css
@@ -23,13 +23,17 @@ button:hover {
 }
 
 .a11y-contrast a11y-contrast-button button,
-.a11y-contrast a11y-font-size-button button {
+.a11y-contrast a11y-font-size-button button,
+.a11y-contrast a11y-font-size-increase-button button,
+.a11y-contrast a11y-font-size-decrease-button button {
   border-color: #e3ff17;
   color: #e3ff17;
 }
 
 .a11y-contrast a11y-contrast-button button:hover,
-.a11y-contrast a11y-font-size-button button:hover {
+.a11y-contrast a11y-font-size-button button:hover,
+.a11y-contrast a11y-font-size-increase-button button:hover,
+.a11y-contrast a11y-font-size-decrease-button button:hover {
   background: #e3ff17;
   color: #000;
 }
